QR-Code-Generator-Scanner/requirements.txt
qrcode[pil]==7.4.2
Pillow==10.0.1
pyzbar==0.1.9
